[Fix] Fix UT to be compatible with pytorch 1.6 (#8707)
* Update

* Update

* force reinstall pycocotools

* Fix build_cuda

* docker install git

* Update

* comment other job to speedup process

* update

* uncomment

* Update

* Update

* Add comments for --force-reinstall

+++ mmdet/datasets/transforms/transforms.py
@@ -1136,7 +1136,7 @@ class MinIoURandomCrop(BaseTransform):
                 if patch[2] == patch[0] or patch[3] == patch[1]:
                     continue
                 overlaps = boxes.overlaps(
-                    HorizontalBoxes(patch.reshape(-1, 4)),
+                    HorizontalBoxes(patch.reshape(-1, 4).astype(np.float32)),
                     boxes).numpy().reshape(-1)
                 if len(overlaps) > 0 and overlaps.min() < min_iou:
                     continue
